Flight Choice Prediction (#31)
Adding the support for the Flight Choice Prediction API & fixing the POST method of the SDK

rb-readline is needed to use guard. guard a library allowing to automatically run all the unit tests and code coverage while working on the project.

One test has been commented: this test was created using vcr, a framework to record test suit. As this test was wrongly recorded, there is no need to keep it.